Azimuth, Tilt, Swing Fixed
No changes in actual code...just comments. Renamed yaw, pitch, roll azimuth, swing, tilt to agree with existing semantics. Also, do not want the user to be confused with yaw pitch roll of air craft.  In addition,  the sign convention of roll (now swing) was described wrong. A CCW rotation of the camera looking from the back of the camera will provide a positive swing.