Populated Folders
GitHub does not like empty folders and will not publish them. So a lot of the saving directories break when they are not present. So the user does not have to create them I prepopulated them with data. Also, this way the user can try different functions without having to run through the others if they miss part of the bootcamp.